Исправляет одинаковые имена инфекшенов на картах
Список ошибок исправленных ошибок:
E     DynamicScene.cpp[1589]22/09 20:50:25 Attempting to load object E_InfectionZone25 of prototype InfectionZone, but an object with this name already exists!
E     DynamicScene.cpp[1589]22/09 20:50:25 Attempting to load object E_InfectionZone90 of prototype InfectionZone, but an object with this name already exists!
E     DynamicScene.cpp[1589]22/09 20:50:25 Attempting to load object E_InfectionZone82 of prototype InfectionZone, but an object with this name already exists!
E     DynamicScene.cpp[1589]22/09 20:50:25 Attempting to load object E_InfectionZone6 of prototype InfectionZone, but an object with this name already exists!
E     DynamicScene.cpp[1589]22/09 20:50:25 Attempting to load object E_InfectionZone22 of prototype InfectionZone, but an object with this name already exists!
E     DynamicScene.cpp[1589]22/09 20:50:25 Attempting to load object E_InfectionZone25 of prototype InfectionZone, but an object with this name already exists!
E     DynamicScene.cpp[1589]22/09 21:00:02 Attempting to load object E_InfectionZone76 of prototype InfectionZone, but an object with this name already exists!
E     DynamicScene.cpp[1589]22/09 21:00:02 Attempting to load object E_InfectionZone21 of prototype InfectionZone, but an object with this name already exists!
E     DynamicScene.cpp[1589]22/09 21:00:02 Attempting to load object E_InfectionZone40 of prototype InfectionZone, but an object with this name already exists!
E     DynamicScene.cpp[1589]22/09 21:00:02 Attempting to load object E_InfectionZone91 of prototype InfectionZone, but an object with this name already exists!
E     DynamicScene.cpp[1589]22/09 21:03:10 Attempting to load object E_InfectionZone72 of prototype InfectionZone, but an object with this name already exists!
E     DynamicScene.cpp[1589]22/09 21:03:10 Attempting to load object E_InfectionZone48 of prototype InfectionZone, but an object with this name already exists!
E     DynamicScene.cpp[1589]22/09 21:03:10 Attempting to load object E_InfectionZone48 of prototype InfectionZone, but an object with this name already exists!
E     DynamicScene.cpp[1589]22/09 21:03:10 Attempting to load object E_InfectionZone19 of prototype InfectionZone, but an object with this name already exists!
E     DynamicScene.cpp[1589]22/09 21:03:10 Attempting to load object E_InfectionZone5 of prototype InfectionZone, but an object with this name already exists!
E     DynamicScene.cpp[1589]22/09 21:03:10 Attempting to load object E_InfectionZone5 of prototype InfectionZone, but an object with this name already exists!
E     DynamicScene.cpp[1589]22/09 21:03:10 Attempting to load object E_InfectionZone13 of prototype InfectionZone, but an object with this name already exists!
E     DynamicScene.cpp[1589]22/09 21:03:10 Attempting to load object E_InfectionZone31 of prototype InfectionZone, but an object with this name already exists!
E     DynamicScene.cpp[1589]22/09 21:03:10 Attempting to load object E_InfectionZone80 of prototype InfectionZone, but an object with this name already exists!
E     DynamicScene.cpp[1589]22/09 21:03:10 Attempting to load object E_InfectionZone46 of prototype InfectionZone, but an object with this name already exists!
E     DynamicScene.cpp[1589]22/09 21:03:10 Attempting to load object E_InfectionZone59 of prototype InfectionZone, but an object with this name already exists!
E     DynamicScene.cpp[1589]22/09 21:03:10 Attempting to load object E_InfectionZone71 of prototype InfectionZone, but an object with this name already exists!
E     DynamicScene.cpp[1589]22/09 21:03:10 Attempting to load object E_InfectionZone74 of prototype InfectionZone, but an object with this name already exists!
E     DynamicScene.cpp[1589]22/09 21:03:10 Attempting to load object E_InfectionZone19 of prototype InfectionZone, but an object with this name already exists!
E     DynamicScene.cpp[1589]22/09 21:03:10 Attempting to load object E_InfectionZone59 of prototype InfectionZone, but an object with this name already exists!
E     DynamicScene.cpp[1589]22/09 21:03:10 Attempting to load object E_InfectionZone75 of prototype InfectionZone, but an object with this name already exists!
E     DynamicScene.cpp[1589]22/09 21:03:10 Attempting to load object E_InfectionZone38 of prototype InfectionZone, but an object with this name already exists!